+++ components/channel/body/ChannelBody.tsx
@@ -1,10 +1,11 @@
 import {cn} from '@/lib/utils';
 
-import {useRef, useEffect, Dispatch, SetStateAction} from 'react';
+import {useRef, useEffect, Dispatch, SetStateAction, useCallback} from 'react';
 
 import {ChannelHistoryType} from '@/types/channel';
 import ChatMessage from '@/components/channel/body/ChatCard';
 import ScrollableContainer from '../../container/ScrollableContainer';
+import useChatSocket from '@/hooks/useChatSocket';
 
 export function ChannelBody({
   messages,
@@ -20,10 +21,36 @@ export function ChannelBody({
   channel_name: string;
 }) {
   const messageEndRef = useRef<HTMLDivElement>();
-
+  const [socket] = useChatSocket('channel');
+  const newMessageHandler = useCallback(
+    ({channelId, userId, userName, profileImage, content}: any) => {
+      console.log('newMessage');
+      console.log(channelId, userId, userName, profileImage, content);
+      console.log('myChannelId', nowChannelId);
+      // if (nowChannelId === channelId) {
+      console.log('메세지가 도착했습니다.');
+      setMessages((prevMessages) => [
+        ...prevMessages,
+        {
+          id: userId,
+          name: userName,
+          profileImage: profileImage,
+          content: content
+        }
+      ]);
+      // }
+    },
+    []
+  );
   useEffect(() => {
     messageEndRef.current?.scrollIntoView({behavior: 'smooth'});
   }, [messages]);
+  useEffect(() => {
+    socket.on('newMessage', newMessageHandler);
+    return () => {
+      socket.off('newMessage', newMessageHandler);
+    };
+  }, []);
   return (
     <div className='h-full'>
       <ScrollableContainer className='rounded-none'>
